src/requirements.txt
transformers>=4.43
accelerate>=0.33
torch>=2.2
datasets>=2.19
pyyaml>=6.0
pandas>=2.2
numpy>=1.26
scipy>=1.13
matplotlib>=3.8
